Spline: always use double for coefficients
Regardless of what's being passed in and out, we want everything
backed by double-precision math.

Don't think this commit is entirely done. I did this while
worrying about something else and then it wasn't important
for that, but I'm putting it here in case it saves me doing
the same work again later.

+++ python/pfs/drp/stella/synthetic.py
@@ -216,12 +216,12 @@ def makeSyntheticDetectorMap(config, minWl=400.0, maxWl=950.0):
     """
     bbox = lsst.geom.Box2I(lsst.geom.Point2I(0, 0), config.dims)
     fiberId = config.fiberId
-    knots = np.arange(config.height, dtype=np.float32)
+    knots = np.arange(config.height, dtype=float)
     xCenter = []
     wavelength = []
     for ii in range(config.numFibers):
-        xCenter.append((config.traceCenters[ii] + config.traceOffset).astype(np.float32))
-        wavelength.append(np.linspace(minWl, maxWl, config.height, dtype=np.float32))
+        xCenter.append((config.traceCenters[ii] + config.traceOffset).astype(float))
+        wavelength.append(np.linspace(minWl, maxWl, config.height, dtype=float))
     return SplinedDetectorMap(bbox, fiberId, [knots]*config.numFibers, xCenter,
                               [knots]*config.numFibers, wavelength)
